Changed basic-usage to angular-app-with-router
