+++ samples/angular-app-with-router/.editorconfig
