+++ samples/angular-app-with-router/.eslintrc.js
