+++ samples/angular-app-with-router/e2e/protractor.conf.js
