+++ samples/angular-app-with-router/e2e/src/app.e2e-spec.ts
@@ -17,26 +17,26 @@
  *
  */
 
-import { Component } from "@angular/core";
-import { AsgardeoAuthService } from "@asgardeo/auth-angular";
+import { browser, logging } from "protractor";
+import { AppPage } from "./app.po";
 
-@Component({
-    selector: "app-home",
-    templateUrl: "./home.component.html",
-    styleUrls: ["./home.component.css"]
-})
-export class HomeComponent {
-    isAuthenticated: boolean;
+describe("workspace-project App", () => {
+    let page: AppPage;
 
-    constructor(private auth: AsgardeoAuthService) {
-        this.isAuthenticated = this.auth.isAuthenticated();
-    }
+    beforeEach(() => {
+        page = new AppPage();
+    });
 
-    signIn(): void {
-        this.auth.signInWithRedirect();
-    }
+    // it("should display welcome message", () => {
+    //     page.navigateTo();
+    //     expect(page.getTitleText()).toEqual("");
+    // });
 
-    signOut(): void {
-        this.auth.signOut();
-    }
-}
+    afterEach(async () => {
+        // Assert that there are no errors emitted from the browser
+        const logs = await browser.manage().logs().get(logging.Type.BROWSER);
+        expect(logs).not.toContain(jasmine.objectContaining({
+            level: logging.Level.SEVERE,
+        } as logging.Entry));
+    });
+});
