+++ samples/angular-app-with-router/e2e/tsconfig.json
