+++ samples/angular-app-with-router/karma.conf.js
@@ -19,7 +19,7 @@
 // Karma configuration file, see link for more information
 // https://karma-runner.github.io/1.0/config/configuration-file.html
 
-module.exports = function(config) {
+module.exports = function (config) {
     config.set({
         basePath: "",
         frameworks: ["jasmine", "@angular-devkit/build-angular"],
@@ -34,7 +34,7 @@ module.exports = function(config) {
             clearContext: false // leave Jasmine Spec Runner output visible in browser
         },
         coverageIstanbulReporter: {
-            dir: require("path").join(__dirname, "./coverage/basic-usage"),
+            dir: require("path").join(__dirname, "./coverage/angular-app-with-router"),
             reports: ["html", "lcovonly", "text-summary"],
             fixWebpackSourcePaths: true
         },
