+++ samples/angular-app-with-router/src/app/app-routing.module.ts
