+++ samples/angular-app-with-router/src/app/app.component.css
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -14,21 +14,4 @@
  * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
- *
  */
-
-div.main-container {
-    padding: 0%!important;
-}
-
-.masthead {
-    margin-bottom: 2rem;
-}
-
-.navbar-brand {
-    margin-left: 1%;
-}
-
-.mast-foot {
-    color: rgba(255, 255, 255, .5);
-}
