+++ samples/angular-app-with-router/src/app/app.component.html
@@ -1,0 +1,51 @@
+<!--
+Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+
+WSO2 Inc. licenses this file to you under the Apache License,
+Version 2.0 (the "License"); you may not use this file except
+in compliance with the License.
+You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied. See the License for the
+specific language governing permissions and limitations
+under the License.
+-->
+
+<img
+    src="/assets/asgardeo-logo.png"
+    class="logo-image"
+/>
+<div class="container">
+    <div *ngIf="isConfigured; then configured else notConfigured"></div>
+
+    <ng-template #notConfigured>
+        <div class="content">
+            <h2>You need to update the Client ID to proceed.</h2>
+            <p>Please open "src/config.json" file using an editor, and update the <code>clientID</code> value with the
+                registered application's client ID.</p>
+            <p>Visit repo <a href="https://github.com/asgardeo/asgardeo-auth-angular-sdk/blob/main/README.md">README</a>
+                for more details.</p>
+        </div>
+    </ng-template>
+
+
+    <ng-template #configured>
+        <div class="header-title">
+            <h1>
+                Angular SPA Authentication Sample - With Router <br /> (OIDC - Authorization Code Grant)
+            </h1>
+        </div>
+
+        <router-outlet></router-outlet>
+
+    </ng-template>
+</div>
+<img
+    src="/assets/footer.png"
+    class="footer-image"
+/>
