+++ samples/angular-app-with-router/src/app/app.module.ts
@@ -20,6 +20,7 @@
 import { NgModule } from "@angular/core";
 import { BrowserModule } from "@angular/platform-browser";
 import { AsgardeoAuthModule, AsgardeoConfigInterface } from "@asgardeo/auth-angular";
+import { NgxJsonViewerModule } from "ngx-json-viewer";
 import { default as authConfig } from "../config.json";
 import { AppRoutingModule } from "./app-routing.module";
 import { AppComponent } from "./app.component";
@@ -35,6 +36,7 @@ import { ProfileComponent } from "./profile/profile.component";
     imports: [
         BrowserModule,
         AppRoutingModule,
+        NgxJsonViewerModule,
         AsgardeoAuthModule.forRoot(authConfig as AsgardeoConfigInterface)
     ],
     providers: [],
