+++ samples/angular-app-with-router/src/config.json
@@ -1,7 +1,7 @@
 {
-    "clientID": "",
+    "clientID": "sjLkiX3wBScX8PwAhilrmCvt5Foa",
     "signInRedirectURL": "https://localhost:5000/signin/redirect",
     "signOutRedirectURL": "https://localhost:5000",
-    "serverOrigin": "",
+    "serverOrigin": "https://dev.portal.asgardio.io/t/angular.test",
     "enablePKCE": true
 }
