+++ samples/angular-app-with-router/src/environments/environment.ts
