+++ samples/angular-app-with-router/src/index.html
