+++ samples/angular-app-with-router/src/main.ts
