+++ samples/angular-app-with-router/src/styles.css
@@ -1,0 +1,189 @@
+/**
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ *
+ */
+
+/* You can add global styles to this file, and also import other style files */
+
+html, body {
+    height: 100vh;
+}
+
+body {
+    margin: 0;
+    padding: 0;
+    text-align: center;
+    color: #2A2A2A;
+    background-color: #fff;
+    font-family: BlinkMacSystemFont, Segoe WPC, Segoe UI, HelveticaNeue-Light, Ubuntu, Droid Sans, sans-serif, Helvetica Neue, Arial, Helvetica, sans-serif;
+}
+
+h1, h2, h3, h4, h5 {
+    font-weight: 300;
+}
+
+h1 b, h2 b, h3 b, h4 b, h5 b {
+    font-weight: 400;
+}
+
+b {
+    font-weight: 500;
+}
+
+a {
+    color: #f47421;
+}
+
+a:hover {
+    text-decoration: none;
+    ;
+}
+
+.mb-0 {
+    margin-bottom: 0 !important;
+}
+
+.mt-4 {
+    margin-top: 40px !important;
+}
+
+.home-image {
+    vertical-align: middle;
+    margin-bottom: 50px;
+}
+
+.home-image .logo {
+    width: 200px;
+    vertical-align: middle;
+}
+
+.home-image .logo-plus {
+    display: inline-block;
+    font-size: 50px;
+    margin: 0 15px 0 30px;
+    vertical-align: middle;
+}
+
+img.footer-image {
+    width: 50px;
+    margin: 30px auto;
+    opacity: .8;
+}
+
+img.logo-image {
+    width: 200px;
+    margin-top: 100px;
+}
+
+code {
+    background: #272822;
+    padding: 4px;
+    font-size: 90%;
+    color: #fd971f;
+}
+
+.code {
+    background: #272822;
+    display: block;
+    word-break: break-all;
+    padding: 20px;
+    text-align: left;
+}
+
+.code .id-token-0 {
+    color: #cc6633;
+}
+
+.code .id-token-1 {
+    color: #f9f8f5;
+}
+
+.code .id-token-2 {
+    color: #fd971f;
+}
+
+.row {
+    width: 100%;
+    display: flex;
+    text-align: left;
+}
+
+.row>.column {
+    flex: 40%;
+    margin-right: 15px;
+}
+
+.row>.column:last-child {
+    flex: 60%;
+    margin-right: 0;
+}
+
+@media screen and (max-width: 600px) {
+    .column {
+        width: 100%;
+        margin-right: 0;
+    }
+}
+
+.container {
+    background: #fff;
+    border: 1px solid #e2e2e2;
+    width: 980px;
+    margin: 0 auto;
+    margin-top: 50px;
+    border-radius: 10px;
+    padding: 0%!important;
+}
+
+.json {
+    text-align: left;
+    display: block;
+    overflow: auto;
+    word-break: break-all;
+}
+
+.json .pretty-json-container.object-container {
+    padding: 20px;
+}
+
+.header-title {
+    background-color: #f47421;
+    color: #ffffff;
+    padding: 20px;
+    border-radius: 10px 10px 0 0;
+}
+
+.content {
+    padding: 50px;
+}
+
+.btn.primary, .btn.primary:active, .btn.primary:focus {
+    background-color: #282c34;
+    border-radius: 3px;
+    border: none;
+    color: #ffffff;
+    display: inline-block;
+    font-size: 1.14285714rem;
+    text-align: center;
+    text-decoration: none;
+    width: 150px;
+    cursor: pointer;
+    -webkit-text-size-adjust: none;
+    padding: .78571429em 1.5em .78571429em;
+    margin-top: 30px;
+    outline: none;
+}
