+++ samples/angular-app-with-router/tsconfig.app.json
