+++ samples/angular-app-with-router/tsconfig.json
