+++ samples/angular-app-with-router/tsconfig.spec.json
