+++ samples/angular-app/e2e/src/app.e2e-spec.ts
@@ -27,10 +27,10 @@ describe("workspace-project App", () => {
         page = new AppPage();
     });
 
-    it("should display welcome message", () => {
-        page.navigateTo();
-        expect(page.getTitleText()).toEqual("");
-    });
+    // it("should display welcome message", () => {
+    //     page.navigateTo();
+    //     expect(page.getTitleText()).toEqual("");
+    // });
 
     afterEach(async () => {
         // Assert that there are no errors emitted from the browser
