--- samples/basic-usage/src/app/app.component.html
@@ -1,40 +1,0 @@
-<!--
-Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-
-WSO2 Inc. licenses this file to you under the Apache License,
-Version 2.0 (the "License"); you may not use this file except
-in compliance with the License.
-You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied. See the License for the
-specific language governing permissions and limitations
-under the License.
--->
-
-<div class="main-container d-flex w-100 h-100 p-3 mx-auto flex-column">
-    <header class="masthead mb-auto">
-        <div class="inner">
-            <nav class="navbar navbar-expand-lg navbar-dark bg-dark sticky-top">
-                <a class="navbar-brand" href="https://wso2.com" target="_blank" rel="noopener noreferrer"><img src="https://wso2.cachefly.net/wso2/sites/all/2020-theme/images/wso2-logo.svg" alt="WSO2 Logo" width="100px" class="ls-is-cached lazyloaded"></a>
-                <h2><a routerLink="" style="color:white;text-decoration: none;">Basic Usage</a></h2>
-            </nav>
-        </div>
-    </header>
-
-    <main role="main" class="inner">
-        <div>
-            <router-outlet></router-outlet>
-        </div>
-    </main>
-
-    <footer class="mast-foot mt-auto">
-        <div class="inner">
-            <p>@2020 <a href="https://wso2.com" target="_blank" rel="noopener noreferrer">WSO2</a></p>
-        </div>
-    </footer>
-</div>
