--- samples/basic-usage/src/app/profile/profile.component.html
@@ -1,42 +1,0 @@
-<!--
- Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-
- WSO2 Inc. licenses this file to you under the Apache License,
- Version 2.0 (the "License"); you may not use this file except
- in compliance with the License.
- You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing,
- software distributed under the License is distributed on an
- "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- KIND, either express or implied. See the License for the
- specific language governing permissions and limitations
- under the License.
-
--->
-
-<div>
-    <div *ngIf="isAuthenticated; then signedIn else notSignedIn"></div>
-
-    <ng-template #notSignedIn>
-        <h3>You are not signed in!!!</h3>
-    </ng-template>
-
-    <ng-template #signedIn>
-
-        <div class="card text-white bg-dark d-flex mx-auto" style="width: fit-content;">
-            <div class="card-header">
-                <h3>{{userInfo.displayName}}</h3>
-            </div>
-            <div class="card-body" style="text-align: left;">
-                <li>Access Token - {{accessToken}}</li>
-                <!-- <li>Email - {{userInfo.email}}</li> -->
-                <li>Username - {{userInfo.username}}</li>
-                <li>Allowed Scopes - {{userInfo.allowedScopes}}</li>
-                <li>Tenant Domain - {{userInfo.tenantDomain}}</li>
-            </div>
-        </div>
-    </ng-template>
-</div>
