--- samples/basic-usage/src/styles.css
@@ -1,51 +1,0 @@
-/**
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-/* You can add global styles to this file, and also import other style files */
-
-@import url("https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css");
-html, body {
-    height: 100vh;
-}
-
-body {
-    align-items: center;
-    background-color: #212a32;
-    background-position: center;
-    display: -ms-flexbox;
-    display: -webkit-box;
-    display: flex;
-    font-family: "Roboto", sans-serif !important;
-    justify-content: center;
-    text-align: center;
-    color: white;
-}
-
-app-root {
-    height: 100%;
-    width: 100%;
-}
-
-@font-face {
-    font-display: swap;
-    font-family: "Roboto";
-    font-style: normal;
-    font-weight: 500;
-    src: url(https://fonts.gstatic.com/s/roboto/v20/KFOlCnqEu92Fr1MmEU9fBBc4.woff2) format("woff2");
-}
